// src/app/shared/error-codes.enum.ts
export enum ErrorCodes {
  AuthErrorIncorrectCredentials = 'AuthErrorIncorrectCredentials',
  AuthErrorEmailAlreadyExists = 'AuthErrorEmailAlreadyExists',
  AuthErrorNotAuthorized = 'AuthErrorNotAuthorized',

  // USER ERRORS
  UserNotFound = 'UserNotFound',
  UserNotUpdated = 'UserNot',
  UserErrorWhenUpdatingUSer = 'UserErrorWhenUpdatingUSe',
  UserErrorWhenCreatingUser = 'UserErrorWhenCreatingUser',
  UserErrorUserNotCreated = 'UserErrorUserNotCreated',
}